fix(*): Standardize element paths
For example, `ons-tabbar/index.js` has been moved to `ons-tabbar.js`.
