+++ esm/elements/ons-popover.js
@@ -15,16 +15,16 @@ limitations under the License.
 
 */
 
-import onsElements from '../../ons/elements.js';
-import util from '../../ons/util.js';
-import autoStyle from '../../ons/autostyle.js';
-import ModifierUtil from '../../ons/internal/modifier-util.js';
-import AnimatorFactory from '../../ons/internal/animator-factory.js';
-import { PopoverAnimator, IOSFadePopoverAnimator, MDFadePopoverAnimator } from './animator.js';
-import platform from '../../ons/platform.js';
-import iPhoneXPatch from '../../ons/iphonex-patch.js';
-import BaseDialogElement from '../base/base-dialog.js';
-import contentReady from '../../ons/content-ready.js';
+import onsElements from '../ons/elements.js';
+import util from '../ons/util.js';
+import autoStyle from '../ons/autostyle.js';
+import ModifierUtil from '../ons/internal/modifier-util.js';
+import AnimatorFactory from '../ons/internal/animator-factory.js';
+import { PopoverAnimator, IOSFadePopoverAnimator, MDFadePopoverAnimator } from './ons-popover/animator.js';
+import platform from '../ons/platform.js';
+import iPhoneXPatch from '../ons/iphonex-patch.js';
+import BaseDialogElement from './base/base-dialog.js';
+import contentReady from '../ons/content-ready.js';
 
 const scheme = {
   '.popover': 'popover--*',
