+++ esm/elements/ons-tabbar.js
@@ -15,14 +15,14 @@ limitations under the License.
 
 */
 
-import onsElements from '../../ons/elements.js';
-import util from '../../ons/util.js';
-import internal from '../../ons/internal/index.js';
-import autoStyle from '../../ons/autostyle.js';
-import Swiper from '../../ons/internal/swiper.js';
-import ModifierUtil from '../../ons/internal/modifier-util.js';
-import BaseElement from '../base/base-element.js';
-import contentReady from '../../ons/content-ready.js';
+import onsElements from '../ons/elements.js';
+import util from '../ons/util.js';
+import internal from '../ons/internal/index.js';
+import autoStyle from '../ons/autostyle.js';
+import Swiper from '../ons/internal/swiper.js';
+import ModifierUtil from '../ons/internal/modifier-util.js';
+import BaseElement from './base/base-element.js';
+import contentReady from '../ons/content-ready.js';
 
 const scheme = {
   '.tabbar__content': 'tabbar--*__content',
